Add files via upload
HTML/ JS code that lets the user choose a CSV file (I've been working with the "scanner data.csv" file for testing purposes that unfortunately has dates saved in the day/month/year format and not month/day/year), and does basic analysis of the mean, mode, and median of the data. You can choose between creating a line or bar chart, and then when you choose  a column to analyze it does those calculations and outputs the chart of your choosing to the main page.